# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "UPA-GURU Public REST API",
    "description": "# UPA-GURU Competitive Exam & Notification API\nThe UPA-GURU Public REST API allows third-party developers, academic portals, and career services\nto programmatically query real-time, verified government competitive exam notifications,\napplication deadlines, vacancy metrics, and official authority metadata.\n\n### Key Features\n- **Standardized Payloads**: Strict JSON contracts across all endpoints.\n- **Rate Limiting**: Sliding window rate limiter enforcing 60 requests/minute per IP address.\n- **RFC 7807 Error Handling**: Granular, machine-readable Problem Details (`application/problem+json`) on validation and runtime failures.\n- **Edge CDN Caching**: Pre-cached at the edge with dynamic `Cache-Control` headers for sub-50ms latency.\n",
    "version": "1.0.0",
    "contact": {
      "name": "UPA-GURU Engineering Team",
      "url": "https://upaguru.in",
      "email": "api-support@upaguru.in"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "https://upaguru.in",
      "description": "Production API Server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development Environment"
    }
  ],
...
  "paths": {
...
    "/api/v1/exams": {
      "get": {
        "tags": [
          "Exams"
        ],
        "summary": "List Competitive Exams",
        "description": "Returns a paginated catalog of all canonical competitive exam entities (e.g., UPSC Civil Services,\nKPSC Gazetted Probationers, SSC CGL) along with conducting authority details and active notification counts.\n",
        "operationId": "getExams",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of exam records to return per page.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of exam records to skip.",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "category",
            "in": "query",
            "description": "Filter exams by category.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "all",
                "civil_services",
                "banking",
                "railways",
                "defense",
                "state_psc",
                "teaching",
                "police",
                "other"
              ],
              "default": "all"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "Filter by state or central jurisdiction.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100,
              "default": "all"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Order of returned exams.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "title_asc",
                "title_desc",
                "created_desc",
                "created_asc"
              ],
              "default": "title_asc"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search term matching exam title, conducting authority, or slug.",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100,
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved exams catalog.",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "example": "public, s-maxage=600, stale-while-revalidate=1200"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "PaginationMetadata": {
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "has_more"
        ],
        "properties": {
          "total": {
            "type": "integer",
            "example": 142,
            "description": "Total number of matching records across all pages."
          },
          "limit": {
            "type": "integer",
            "example": 20,
            "description": "Maximum number of records requested."
          },
          "offset": {
            "type": "integer",
            "example": 0,
            "description": "Number of records skipped."
          },
          "has_more": {
            "type": "boolean",
            "example": true,
            "description": "Whether additional pages exist beyond the current offset + limit."
          }
        }
      },
...
      "ExamItem": {
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "conducting_body",
          "category",
          "state_or_central",
          "official_website",
          "active_notifications_count",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "4955b252-47ba-4b21-a1e4-39958ee42f21"
          },
          "slug": {
            "type": "string",
            "example": "kpsc-gazetted-probationers"
          },
          "title": {
            "type": "string",
            "example": "Karnataka Administrative Services (KAS)"
          },
          "conducting_body": {
            "type": "string",
            "example": "Karnataka Public Service Commission"
          },
          "category": {
            "type": "string",
            "example": "state_psc"
          },
          "state_or_central": {
            "type": "string",
            "example": "Karnataka"
          },
          "official_website": {
            "type": "string",
            "format": "uri",
            "example": "https://kpsc.kar.nic.in"
          },
          "logo_url": {
            "type": "string",
            "format": "uri",
            "nullable": true,
            "example": "https://upaguru.in/logos/kpsc.png"
          },
          "active_notifications_count": {
            "type": "integer",
            "example": 3,
            "description": "Number of active published notifications currently associated with this exam."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-01-01T00:00:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-01-01T00:00:00Z"
          }
        }
      },
      "ExamsResponse": {
        "type": "object",
        "required": [
          "data",
          "pagination",
          "timestamp"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExamItem"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMetadata"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-20T12:00:00.000Z"
          }
        }
      },
...
      "InvalidParam": {
        "type": "object",
        "required": [
          "name",
          "reason"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "limit"
          },
          "reason": {
            "type": "string",
            "example": "Limit cannot exceed 100"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "example": "https://upaguru.in/errors/invalid-parameters"
          },
          "title": {
            "type": "string",
            "example": "Invalid Query Parameters"
          },
          "status": {
            "type": "integer",
            "example": 400
          },
          "detail": {
            "type": "string",
            "example": "The request query parameters failed validation: limit (Limit cannot exceed 100)"
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/notifications"
          },
          "invalid_params": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvalidParam"
            }
          }
        }
      }
    }
  }
}